Refactor roles :wrench:
