+++ client/admin/router.js
@@ -1,6 +1,6 @@
 import get from 'lodash/get';
 import NotFound from '@/admin/components/common/NotFound';
-import role from '@/../common/config/role';
+import { Role } from '@/../common/config';
 import Router from 'vue-router';
 import store from './store';
 import Users from '@/admin/components/users';
@@ -23,7 +23,7 @@ const router = new Router({
 router.beforeEach((to, from, next) => {
   const user = get(store.state, 'auth.user');
   const isNotAuthenticated = to.matched.some(it => it.meta.auth) && !user;
-  const isNotAuthorized = user && user.role !== role.ADMIN;
+  const isNotAuthorized = user && user.role !== Role.Admin;
   if (isNotAuthenticated || isNotAuthorized) return loadMainSpa();
   next();
 });
