+++ client/main/router.js
@@ -5,7 +5,7 @@ import Home from '@/main/components';
 import Login from '@/main/components/auth/Login';
 import NotFound from '@/admin/components/common/NotFound';
 import ResetPassword from '@/main/components/auth/ResetPassword';
-import role from '@/../common/config/role';
+import { Role } from '@/../common/config';
 import Router from 'vue-router';
 import store from './store';
 import Vue from 'vue';
@@ -44,7 +44,7 @@ router.beforeEach((to, from, next) => {
   const user = get(store.state, 'auth.user');
   const isNotAuthenticated = to.matched.some(it => it.meta.auth) && !user;
   if (isNotAuthenticated) return next({ name: 'login' });
-  if (user && user.role === role.ADMIN) {
+  if (user && user.role === Role.Admin) {
     document.location.replace(`${document.location.origin}/admin`);
   }
   return next();
